moving coloring tests inside the module
